Develop

+++ modules/staffshares/staffshares-api/src/main/java/com/bjike/goddess/staffshares/bo/DetailsBO.java
@@ -31,7 +31,7 @@ public class DetailsBO extends BaseBO {
     /**
      * 出售/发行数量
      */
-    private int number;
+    private Long number;
 
     /**
      * 出售/发行价格
@@ -46,7 +46,7 @@ public class DetailsBO extends BaseBO {
     /**
      * 剩余出售量
      */
-    private int sharesNum;
+    private Long sharesNum;
 
 
     public String getCode() {
@@ -73,11 +73,11 @@ public class DetailsBO extends BaseBO {
         this.publisher = publisher;
     }
 
-    public int getNumber() {
+    public Long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(Long number) {
         this.number = number;
     }
 
@@ -97,11 +97,11 @@ public class DetailsBO extends BaseBO {
         this.time = time;
     }
 
-    public int getSharesNum() {
+    public Long getSharesNum() {
         return sharesNum;
     }
 
-    public void setSharesNum(int sharesNum) {
+    public void setSharesNum(Long sharesNum) {
         this.sharesNum = sharesNum;
     }
 }
