+++ modules/staffshares/staffshares-api/src/main/java/com/bjike/goddess/staffshares/bo/SchemeIssueBO.java
@@ -81,7 +81,7 @@ public class SchemeIssueBO extends BaseBO {
     /**
      * 发行数量
      */
-    private int number;
+    private Long number;
 
     /**
      * 发行价格
@@ -121,7 +121,7 @@ public class SchemeIssueBO extends BaseBO {
     /**
      * 剩余出售股数
      */
-    private int sharesNum;
+    private Long sharesNum;
 
     /**
      * 备注
@@ -209,11 +209,11 @@ public class SchemeIssueBO extends BaseBO {
         this.proportion = proportion;
     }
 
-    public int getNumber() {
+    public Long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(Long number) {
         this.number = number;
     }
 
@@ -273,11 +273,11 @@ public class SchemeIssueBO extends BaseBO {
         this.accountingTime = accountingTime;
     }
 
-    public int getSharesNum() {
+    public Long getSharesNum() {
         return sharesNum;
     }
 
-    public void setSharesNum(int sharesNum) {
+    public void setSharesNum(Long sharesNum) {
         this.sharesNum = sharesNum;
     }
