+++ modules/staffshares/staffshares-api/src/main/java/com/bjike/goddess/staffshares/dto/SellscheduleDTO.java
@@ -30,7 +30,7 @@ public class SellscheduleDTO extends BaseDTO {
     /**
      * 出售股数
      */
-    private int sellNum;
+    private Long sellNum;
 
     /**
      * 出售价格
@@ -50,7 +50,7 @@ public class SellscheduleDTO extends BaseDTO {
     /**
      * 剩余出售量
      */
-    private int number;
+    private Long number;
 
     /**
      * 购买人
@@ -60,7 +60,7 @@ public class SellscheduleDTO extends BaseDTO {
     /**
      * 购买股数
      */
-    private int purchaseNum;
+    private Long purchaseNum;
 
     /**
      * 购买时间
@@ -91,11 +91,11 @@ public class SellscheduleDTO extends BaseDTO {
         this.name = name;
     }
 
-    public int getSellNum() {
+    public Long getSellNum() {
         return sellNum;
     }
 
-    public void setSellNum(int sellNum) {
+    public void setSellNum(Long sellNum) {
         this.sellNum = sellNum;
     }
 
@@ -123,11 +123,11 @@ public class SellscheduleDTO extends BaseDTO {
         this.sellTime = sellTime;
     }
 
-    public int getNumber() {
+    public Long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(Long number) {
         this.number = number;
     }
 
@@ -139,11 +139,11 @@ public class SellscheduleDTO extends BaseDTO {
         this.buyName = buyName;
     }
 
-    public int getPurchaseNum() {
+    public Long getPurchaseNum() {
         return purchaseNum;
     }
 
-    public void setPurchaseNum(int purchaseNum) {
+    public void setPurchaseNum(Long purchaseNum) {
         this.purchaseNum = purchaseNum;
     }
