+++ modules/staffshares/staffshares-api/src/main/java/com/bjike/goddess/staffshares/vo/SellscheduleVO.java
@@ -33,7 +33,7 @@ public class SellscheduleVO {
     /**
      * 出售股数
      */
-    private int sellNum;
+    private Long sellNum;
 
     /**
      * 出售价格
@@ -53,7 +53,7 @@ public class SellscheduleVO {
     /**
      * 剩余出售量
      */
-    private int number;
+    private Long number;
 
     /**
      * 购买人
@@ -63,7 +63,7 @@ public class SellscheduleVO {
     /**
      * 购买股数
      */
-    private int purchaseNum;
+    private Long purchaseNum;
 
     /**
      * 购买时间
@@ -103,11 +103,11 @@ public class SellscheduleVO {
         this.name = name;
     }
 
-    public int getSellNum() {
+    public Long getSellNum() {
         return sellNum;
     }
 
-    public void setSellNum(int sellNum) {
+    public void setSellNum(Long sellNum) {
         this.sellNum = sellNum;
     }
 
@@ -135,11 +135,11 @@ public class SellscheduleVO {
         this.sellTime = sellTime;
     }
 
-    public int getNumber() {
+    public Long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(Long number) {
         this.number = number;
     }
 
@@ -151,11 +151,11 @@ public class SellscheduleVO {
         this.buyName = buyName;
     }
 
-    public int getPurchaseNum() {
+    public Long getPurchaseNum() {
         return purchaseNum;
     }
 
-    public void setPurchaseNum(int purchaseNum) {
+    public void setPurchaseNum(Long purchaseNum) {
         this.purchaseNum = purchaseNum;
     }
